Extract logIn/saveUser from server.js into users.js
Move the user logic verbatim into its own module, leaving the routes as
thin adapters, so it can be tested directly against a real Parse Server
instead of only through HTTP. The characterization tests from the previous
commit stay green through the move.
